Add files via upload
This project is about wrangling data for WeRateDogs account on twitter. The data wrangling process is divided over three stages gathering, assessing, and cleaning
Thankfully I have passed the project on Udacity, so I post it here to share the knowledge and I hope you find my code helpful or at least part of it.
Finally, two reports uploaded descripting the wrangling effort and the other the analysis and visualization approach.